docker_app/requirements.txt
boto3
mcp[cli]
nova-act
opensearch-py
pandas
retrying
strands-agents 
strands-agents-tools[mem0_memory]
streamlit
streamlit-cognito-auth
pytz
schedule
requests
urllib3
